refactor: adjust antv-dagre test case
